chore: minor changes/editions and package installations
