+++ package.json
@@ -22,19 +22,31 @@
 		"@octokit/plugin-create-or-update-text-file": "^4.0.1",
 		"@octokit/plugin-request-log": "^4.0.0",
 		"@octokit/plugin-retry": "^6.0.1",
+		"@radix-ui/react-accordion": "^1.2.0",
+		"@radix-ui/react-checkbox": "^1.1.0",
+		"@radix-ui/react-collapsible": "^1.1.0",
+		"@radix-ui/react-dialog": "^1.1.1",
+		"@radix-ui/react-icons": "^1.3.0",
+		"@radix-ui/react-popover": "^1.1.1",
+		"@radix-ui/react-separator": "^1.1.0",
+		"@radix-ui/react-slot": "^1.1.0",
 		"@t3-oss/env-nextjs": "^0.8.0",
 		"axios": "^1.6.7",
 		"class-variance-authority": "^0.7.0",
 		"clsx": "^2.1.0",
 		"csp-header": "^5.2.1",
 		"framer-motion": "^11.0.5",
+		"geist": "^1.3.0",
 		"jotai": "^2.6.4",
+		"jotai-optics": "^0.4.0",
 		"lucide-react": "^0.323.0",
 		"next": "^14.1.0",
 		"next-themes": "^0.2.1",
 		"octokit": "^3.1.2",
-		"react": "18.2.0",
+		"react": "18.3.1",
 		"react-hook-form": "^7.50.1",
+		"react-resizable-panels": "^2.0.19",
+		"react-use-measure": "^2.1.1",
 		"tailwind-merge": "^2.2.1",
 		"tailwindcss-animate": "^1.0.7",
 		"usehooks-ts": "^2.14.0",
@@ -47,17 +59,15 @@
 		"@faker-js/faker": "^8.4.1",
 		"@playwright/test": "^1.41.2",
 		"@tailwindcss/typography": "^0.5.10",
-		"@types/eslint": "^8.56.2",
+		"@types/eslint": "^8.56.10",
 		"@types/node": "^20.11.19",
 		"@types/prettier": "^3.0.0",
 		"@types/react": "^18.2.57",
 		"@types/react-dom": "^18.2.19",
-		"@types/uuid": "^9.0.8",
 		"@typescript-eslint/eslint-plugin": "^7.0.2",
 		"@typescript-eslint/parser": "^7.0.2",
-		"autoprefixer": "^10.4.17",
-		"dotenv": "^16.4.4",
-		"drizzle-kit": "^0.20.14",
+		"autoprefixer": "^10.4.19",
+		"dotenv": "^16.4.5",
 		"eslint": "^8.54.0",
 		"eslint-config-next": "^14.1.0",
 		"eslint-config-prettier": "^9.0.0",
@@ -66,11 +76,10 @@
 		"husky": "^9.0.11",
 		"is-ci": "^3.0.1",
 		"lint-staged": "^15.2.2",
-		"pg": "^8.11.3",
-		"postcss": "^8.4.35",
+		"postcss": "^8.4.38",
 		"prettier": "^3.2.5",
 		"prettier-plugin-tailwindcss": "^0.5.11",
-		"tailwindcss": "^3.4.1",
+		"tailwindcss": "^3.4.4",
 		"tsx": "^4.7.1",
 		"typescript": "^5.3.2"
 	},
