+++ src/components/editor/menubar/index.tsx
@@ -1,8 +1,2 @@
-export { CopyToClipboard } from "./CopyToClipboard";
-export { MoveToTrash } from "./MoveToTrash";
-export { ToggleTheme } from "./ChangeTheme";
 export { SettingsModal } from "./Settings";
-export { MakeFavorite } from "./Favorite";
-export { DownloadNote } from "./Download";
 export { EditorMenubar } from "./Menubar";
-export { SyncNotes } from "./SyncNotes";
