+++ tsconfig.json
@@ -18,7 +18,7 @@
 		"noUncheckedIndexedAccess": true,
 		"baseUrl": "./",
 		"paths": {
-			"@/*": ["./src/*"],
+			"@/*": ["./src/*", "./src/components/*"],
 		},
 		"plugins": [
 			{
